Decompile chrAdjustPosForSpawn

+++ src/include/lib/lib_17ce0.h
@@ -7,6 +7,6 @@
 u32 func00017ce0(void);
 u32 func00017dc4(void);
 bool func00017e30(s32 portalnum, struct coord *arg1, struct coord *arg2);
-void func00018148(struct coord *pos, struct coord *pos2, s16 *rooms, u32 *arg3, s16 *arg4, s32 arg5);
+void func00018148(struct coord *pos, struct coord *pos2, s16 *rooms, s16 *arg3, s16 *arg4, s32 arg5);
 
 #endif
